# Email regex optimization