--- a/validator_test.go
+++ b/validator_test.go
@@ -8756,6 +8756,12 @@ func TestEmail(t *testing.T) {
 	NotEqual(t, errs, nil)
 	AssertError(t, errs, "", "", "", "", "email")
 
+	// Multiple consecutive dots
+	s = "test@yahoo.co...jp"
+	errs = validate.Var(s, "email")
+	NotEqual(t, errs, nil)
+	AssertError(t, errs, "", "", "", "", "email")
+
 	i := true
 	errs = validate.Var(i, "email")
 	NotEqual(t, errs, nil)