@deankarn 2025-02-15
I like the simplification and it appears correct, my only FUD is, how confident are we that it functions the same as the previous?

I mean the test still all pass etc.. but wanted to gauge you confidence level as well @nodivbyzero