@nodivbyzero 2025-02-17
@deankarn 
I'm pretty confident in this regex. Honestly, I used ChatGPT to optimize the original regex and then tested it with different email addresses. 
All tests still pass, so I believe it functions the same as before.